# cmd/docs/swagger.yaml
definitions:
  models.AddressInfo:
    properties:
      city:
        type: string
      house_name:
        type: string
      name:
        type: string
      pin:
        type: string
      state:
        type: string
      street:
        type: string
    required:
    - house_name
    - name
    - pin
    - state
    type: object
...
  models.ChangePassword:
    properties:
      old_password:
        type: string
      password:
        type: string
      re_password:
        type: string
    type: object
  models.ForgotPasswordSend:
    properties:
      phone:
        type: string
    type: object
  models.ForgotVerify:
    properties:
      newpassword:
        maxLength: 20
        minLength: 6
        type: string
      otp:
        type: string
      phone:
        type: string
    required:
    - newpassword
    - otp
    - phone
    type: object
  models.LoginDetail:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  models.UserSignUp:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  models.UsersProfileDetails:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      phone:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Zhooze is an E-commerce platform to purchase and sell shoes
  title: Go + Gin Zhooze E-Commerce API
  version: 1.0.0
paths:
...
  /user/address:
    delete:
      consumes:
      - application/json
      description: Delete From User Profile
      parameters:
      - description: address id
        in: query
        name: address_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete User Address
      tags:
      - User Profile
    get:
      consumes:
      - application/json
      description: user can get all their addresses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get Addresses
      tags:
      - User Profile
    post:
      consumes:
      - application/json
      description: user can add their addresses
      parameters:
      - description: address
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/models.AddressInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add Address
      tags:
      - User Profile
    put:
      consumes:
      - application/json
      description: Update User address by sending in address id
      parameters:
      - description: address id
        in: query
        name: address_id
        required: true
        type: string
      - description: User Address Input
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/models.AddressInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update User Address
      tags:
      - User Profile
...
  /user/forgot-password:
    post:
      consumes:
      - application/json
      description: user can change their password if user forgot the password and
        login
      parameters:
      - description: forgot-send
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/models.ForgotPasswordSend'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Forgot password Send OTP
      tags:
      - User
  /user/forgot-password-verify:
    post:
      consumes:
      - application/json
      description: user can change their password if user forgot the password and
        login
      parameters:
      - description: forgot-verify
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/models.ForgotVerify'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Forgot password Verfy and Change
      tags:
      - User
...
  /user/signup:
    post:
      consumes:
      - application/json
      description: user can signup by giving their details
      parameters:
      - description: signup
        in: body
        name: signup
        required: true
        schema:
          $ref: '#/definitions/models.UserSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Signup
      tags:
      - User
  /user/userlogin:
    post:
      consumes:
      - application/json
      description: user can log in by giving their details
      parameters:
      - description: login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.LoginDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Login
      tags:
      - User
  /user/users:
    get:
      consumes:
      - application/json
      description: User Details from User Profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: User Details
      tags:
      - User Profile
    put:
      consumes:
      - application/json
      description: Update User Details by sending in user id
      parameters:
      - description: User Details Input
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/models.UsersProfileDetails'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update User Details
      tags:
      - User Profile
  /user/users/changepassword:
    put:
      consumes:
      - application/json
      description: Change User Password
      parameters:
      - description: User Password Change
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Change User Password
      tags:
      - User Profile
...
  /user/wishlist:
    delete:
      consumes:
      - application/json
      description: Delete From wish List
      parameters:
      - description: product id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete From Wishlist
      tags:
      - WishList Management
    get:
      consumes:
      - application/json
      description: Display wish List
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Display Wishlist
      tags:
      - WishList Management
    post:
      consumes:
      - application/json
      description: Add To wish List
      parameters:
      - description: product id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add to Wishlist
      tags:
      - WishList Management
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"